utilities/requirements.txt
pycryptodome ~= 3.18.0
